refactor(phase-2): sdk-client-v2 → ts-client 4, platform-sdk 8 (2.2 part 1)
- @commercetools/sdk-client-v2 (deprecated 10/2024) removed; ts-client 4.10
- platform-sdk 4.11 → 8.27; all 14 deep dist/declarations imports replaced
  with public root exports across 12 files
- BuildClient.ts → ctClient.ts: ts-client v4 API (async TokenCache,
  httpClient option), same four auth flows and factory signatures so
  services/stores stay untouched until their own phases
- vite workarounds for the dead SDK removed: browser-build aliases and
  define.global polyfill
- NodeJS.Timeout in browser code → ReturnType<typeof setTimeout> (was held
  up by the old SDK's transitive node types)
- gates: tsc, eslint 0 errors, 57/57 unit, build 3.8s, e2e 11/11
