+++ package.json
@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@commercetools/platform-sdk": "^4.11.0",
-    "@commercetools/sdk-client-v2": "^2.2.0",
+    "@commercetools/platform-sdk": "^8.27.0",
+    "@commercetools/ts-client": "^4.10.0",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.1",
     "@mui/icons-material": "^5.18.0",
