+++ src/utils/helpers.ts
@@ -1,5 +1,5 @@
 export function debounce<F extends (...args: never[]) => void>(func: F, delay: number): F {
-  let inDebounce: NodeJS.Timeout | null = null;
+  let inDebounce: ReturnType<typeof setTimeout> | null = null;
 
   return function debouncedFunction(...args: Parameters<F>) {
     if (inDebounce !== null) {
